requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.1
requests==2.31.0
gunicorn==21.2.0